refactor: Split instructions module into semantic groups
Reorganize CPU instructions into separate modules by functionality:
- instructions/load_store.rs: LDA, LDX, LDY, STA, STX, STY
- instructions/transfer.rs: TAX, TAY, TXA, TYA, TSX, TXS
- instructions/arithmetic.rs: ADC, SBC, INC, DEC, INX, INY, DEX, DEY
- instructions/mod.rs: Module declarations and helper functions

This structure improves code organization and maintainability as the
instruction set implementation continues to expand. Each file now
contains related instructions with their tests, making it easier to
locate and modify specific instruction groups.

All 120 tests pass after refactoring.
